Fix another ENV issue
emerge was demonstrating occasional failures in resolving external
commands, such as bzip2. This was because of ENV being empty. While
it silently assumed the existance of ENV['path'].

An alternative workaround would have been to just provide a PATH to
that command, but it seemed wiser for the moment to just keep the
original ENV in place.